backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# File handling
python-multipart==0.0.20
pandas==2.2.3
openpyxl==3.1.5

# LLM providers
google-generativeai==0.8.3
groq==0.13.0

# Email
aiosmtplib==3.0.2

# Config & validation
pydantic-settings==2.7.0
pydantic[email]==2.10.4

# Rate limiting
slowapi==0.1.9

# File type detection (magic bytes)
python-magic-bin==0.4.14
